+++ plugins/config/crates/model/src/component/mod.rs
@@ -1,0 +1,3 @@
+pub use config_file::*;
+
+pub mod config_file;