+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - Refactored into new mono repo `reactive-graph/sys`
+- Moved plugins binary, config, file and system-environment from `reactive-graph/std` -> `reactive-graph/sys`
 - Prefix plugins with `sys` (e.g. `libreactive_graph_sys_binary`)
 
 ### Fixed